+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## [DEV] - 2021-04-12
 - updated encore bundle integration
+- fixed contao 4.4 compatibility
+- fixed missing dependencies in composer.json
 
 ## [0.2.1] - 2021-01-07
 - fixed handling of form submit on errors